LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-history-interface/history_go_zero-expected.txt

main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)

FAIL history.go(0) assert_array_equals: Pages opened during history navigation lengths differ, expected 3 got 2
